@@ -15,49 +15,46 @@
     "test:coverage": "jest --coverage",
     "test:watch": "jest --watch",
     "test:ci": "jest --ci",
-    "storybook": "start-storybook -p 6006",
-    "storybook:build": "build-storybook",
+    "storybook": "storybook dev -p 6006",
+    "storybook:build": "storybook build",
     "postinstall": "husky install",
     "generate": "yarn plop --plopfile ./.plop/plopfile.js"
   },
-  "lint-staged": {
-    "src/**/*": [
-      "yarn prettier:format",
-      "yarn eslint:format"
-    ]
-  },
   "dependencies": {
-    "clsx": "1.2.1",
-    "next": "13.4.2",
+    "@t3-oss/env-nextjs": "0.4.0",
+    "add": "2.0.6",
+    "next": "13.4.4",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "tailwind-merge": "1.13.0",
+    "zod": "3.21.4"
   },
   "devDependencies": {
-    "@storybook/addon-a11y": "6.5.16",
-    "@storybook/addon-actions": "6.5.16",
-    "@storybook/addon-essentials": "6.5.16",
-    "@storybook/addon-interactions": "6.5.16",
-    "@storybook/addon-links": "6.5.16",
-    "@storybook/addon-postcss": "2.0.0",
-    "@storybook/addon-storysource": "6.5.16",
-    "@storybook/react": "6.5.16",
+    "@storybook/addon-essentials": "^7.0.18",
+    "@storybook/addon-interactions": "^7.0.18",
+    "@storybook/addon-links": "^7.0.18",
+    "@storybook/blocks": "^7.0.18",
+    "@storybook/nextjs": "^7.0.18",
+    "@storybook/react": "^7.0.18",
+    "@storybook/testing-library": "^0.0.14-next.2",
     "@testing-library/jest-dom": "5.16.5",
     "@testing-library/react": "14.0.0",
     "@testing-library/user-event": "14.4.3",
-    "@types/node": "18.16.3",
-    "@types/react": "18.0.37",
+    "@types/node": "20.2.5",
+    "@types/react": "18.2.8",
     "@types/react-dom": "18.2.4",
-    "@typescript-eslint/eslint-plugin": "5.59.6",
-    "@typescript-eslint/parser": "5.59.6",
+    "@typescript-eslint/eslint-plugin": "5.59.8",
+    "@typescript-eslint/parser": "5.59.8",
     "autoprefixer": "10.4.14",
     "babel-jest": "29.5.0",
-    "eslint": "8.41.0",
+    "eslint": "8.42.0",
     "eslint-config-next": "13.4.4",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-import-helpers": "1.3.1",
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-react": "7.32.2",
     "eslint-plugin-react-hooks": "4.6.0",
+    "eslint-plugin-storybook": "0.6.12",
     "eslint-plugin-testing-library": "5.11.0",
     "husky": "8.0.3",
     "identity-obj-proxy": "3.0.0",
@@ -67,13 +64,18 @@
     "lint-staged": "13.2.2",
     "pinst": "3.0.0",
     "plop": "3.1.2",
-    "postcss": "8.4.23",
+    "postcss": "8.4.24",
     "prettier": "2.8.8",
     "prettier-plugin-tailwindcss": "0.3.0",
-    "storybook-addon-next-router": "4.0.2",
-    "storybook-dark-mode": "3.0.0",
-    "tailwind-scrollbar": "3.0.0",
+    "storybook": "^7.0.18",
+    "tailwind-scrollbar": "3.0.4",
     "tailwindcss": "3.3.2",
-    "typescript": "4.9.5"
+    "typescript": "5.1.3"
+  },
+  "lint-staged": {
+    "src/**/*": [
+      "yarn prettier:format",
+      "yarn eslint:format"
+    ]
   }
 }
