@@ -1,6 +1,8 @@
+import './src/lib/env/env.mjs';
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true
 };
 
-module.exports = nextConfig;
+export default nextConfig;
